+++ common/src/main/java/me/cg360/mod/bridging/mixin/MinecraftClientMixin.java
@@ -105,7 +105,8 @@ public abstract class MinecraftClientMixin {
             int originalStackSize = itemStack.getCount();
             InteractionResult blockPlaceResult = this.gameMode.useItemOn(this.player, hand, blockHitResult);
 
-            if (!blockPlaceResult.consumesAction()) continue;
+
+            if (!(blockPlaceResult instanceof InteractionResult.Success successResult)) continue;
 
             // if successful place occurred, cancel all future behaviour for
             // item placement as this takes over instead. Stops off-hand
@@ -113,8 +114,8 @@ public abstract class MinecraftClientMixin {
             this.rightClickDelay = Math.max(0, BridgingMod.getConfig().getDelayPostBridging());
             info.cancel();
 
-            if(true) return;
-            //if (!blockPlaceResult.shouldSwing()) return;
+            if(successResult.swingSource() != InteractionResult.SwingSource.CLIENT)
+                return;
 
             this.player.swing(hand);
             boolean stackSizeChanged = itemStack.getCount() != originalStackSize || this.gameMode.hasInfiniteItems();